safari-600.7-branch/LayoutTests/platform/ios-sim/webgl/conformance/extensions/webgl-compressed-texture-pvrtc-expected.txt

CONSOLE MESSAGE: line 252: TypeError: undefined is not a function (evaluating 'wtu.clearAndDrawUnitQuad(gl)')
This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance/extensions/webgl-compressed-texture-pvrtc.html
FAIL
